Improve booting and UI handling
Still getting the hang of Dioxus. Everything still looks awful, but
maybe slightly less so than before?

Anyway, the device handling and UI for the boot is taking shape. Most
of the bits are in place to handle taking a device from vendor fastboot
thru to the kernel (via U-Boot).